requirements.txt
discord.py
python-dotenv

Wikipedia-API